FlipX and SetInvincible properly create mixer behaviors, reorganized timeline scripts
