+++ YearLongProject/Assets/Fighters/Ahab/Ahab.prefab
@@ -606,7 +606,7 @@ PlayableDirector:
   m_Enabled: 1
   serializedVersion: 3
   m_PlayableAsset: {fileID: 11400000, guid: 4777c4b8df04be741b0d782f644ff716, type: 2}
-  m_InitialState: 1
+  m_InitialState: 0
   m_WrapMode: 2
   m_DirectorUpdateMode: 1
   m_InitialTime: 0
