+++ YearLongProject/Assets/Fighters/Ahab/Timelines/AhabDash.playable
@@ -599,10 +599,10 @@ MonoBehaviour:
   m_Children: []
   m_Clips:
   - m_Version: 1
-    m_Start: 0.08333333333333333
+    m_Start: 0
     m_ClipIn: 0
     m_Asset: {fileID: 6673962422136390862}
-    m_Duration: 0.16666666666666669
+    m_Duration: 0.20540946225325268
     m_TimeScale: 1
     m_ParentTrack: {fileID: 2670663638000722055}
     m_EaseInDuration: 0
