+++ YearLongProject/Assets/Scripts/Timeline/LockFlipX/LockFlipXPlayableBehavior.cs
@@ -2,28 +2,42 @@ using CharacterScripts;
 using UnityEngine;
 using UnityEngine.Playables;
 
-namespace Timeline
+namespace Timeline.LockFlipX
 {
     public class LockFlipXPlayableBehavior : PlayableBehaviour
     {
-        FlipXHandler flipX;
+        private FlipXHandler flipX;
+
         public override void OnBehaviourPlay(Playable playable, FrameData info)
         {
-            if (!Application.isPlaying) return;
+            if (!Application.isPlaying)
+            {
+                return;
+            }
 
             flipX = info.output.GetUserData() as FlipXHandler;
 
-            if (flipX == null) return;
+            if (flipX == null)
+            {
+                return;
+            }
+
             flipX.CanFlipX = false;
         }
 
         public override void OnBehaviourPause(Playable playable, FrameData info)
         {
-            if (!Application.isPlaying) return;
-            if (flipX == null) return;
+            if (!Application.isPlaying)
+            {
+                return;
+            }
+
+            if (flipX == null)
+            {
+                return;
+            }
 
             flipX.CanFlipX = true;
         }
     }
 }
-
