--- YearLongProject/Assets/Scripts/Timeline/LockFlipXPlayableAsset.cs
@@ -1,18 +1,0 @@
-using UnityEngine;
-using UnityEngine.Playables;
-using UnityEngine.Timeline;
-
-namespace Timeline
-{ 
-    public class LockFlipXPlayableAsset : PlayableAsset, ITimelineClipAsset
-    {
-            public LockFlipXPlayableBehavior template = new();
-
-            public ClipCaps clipCaps => ClipCaps.None;
-
-            public override Playable CreatePlayable(PlayableGraph graph, GameObject owner)
-            {
-                return ScriptPlayable<LockFlipXPlayableBehavior>.Create(graph, template);
-            }
-    }
-}
