--- YearLongProject/Assets/Scripts/Timeline/LockFlipXTrack.cs
@@ -1,18 +1,0 @@
-using CharacterScripts;
-using UnityEngine;
-using UnityEngine.Playables;
-using UnityEngine.Timeline;
-
-namespace Timeline
-{
-    // A track that allows timed restriction of changing directions
-    [TrackBindingType(typeof(FlipXHandler))]
-    [TrackClipType(typeof(LockFlipXPlayableAsset))]
-    public class LockFlipXTrack : TrackAsset
-    {
-        public override Playable CreateTrackMixer(PlayableGraph graph, GameObject go, int inputCount)
-        {
-            return ScriptPlayable<LockFlipXPlayableBehavior>.Create(graph, inputCount);
-        }
-    }
-}
