+++ YearLongProject/Assets/Scripts/Timeline/RigidbodyTween/GravityTween/GravityTweenPlayableBehavior.cs
@@ -48,7 +48,6 @@ namespace Timeline.RigidbodyTween.GravityTween
 
             initialGravity = rigidbody.Gravity;
 
-            Debug.Log(initialGravity);
             base.OnBehaviourPlay(playable, info);
         }
 
@@ -59,12 +58,10 @@ namespace Timeline.RigidbodyTween.GravityTween
                 return;
             }
 
-            
             rigidbody.Gravity = EndGravity;
             if (ReturnToStartGravity)
             {
-                rigidbody.Gravity = initialGravity; 
-                Debug.Log($"resetting to {initialGravity}");
+                rigidbody.Gravity = initialGravity;
             }
         }
     }
