+++ YearLongProject/Assets/Scripts/Timeline/SetInvincible/SetInvinciblePlayableBehavior.cs
@@ -2,26 +2,40 @@ using GameEntities;
 using UnityEngine;
 using UnityEngine.Playables;
 
-namespace Timeline
+namespace Timeline.SetInvincible
 {
     public class SetInvinciblePlayableBehavior : PlayableBehaviour
     {
-        Entity entity;
+        private Entity entity;
+
         public override void OnBehaviourPlay(Playable playable, FrameData info)
         {
-            if (!Application.isPlaying) return;
+            if (!Application.isPlaying)
+            {
+                return;
+            }
 
             entity = info.output.GetUserData() as Entity;
-            
-            if (entity == null) return;
+
+            if (entity == null)
+            {
+                return;
+            }
 
             entity.IsInvincible = true;
         }
 
         public override void OnBehaviourPause(Playable playable, FrameData info)
         {
-            if (!Application.isPlaying) return;
-            if (entity == null) return;
+            if (!Application.isPlaying)
+            {
+                return;
+            }
+
+            if (entity == null)
+            {
+                return;
+            }
 
             entity.IsInvincible = false;
         }
