+++ YearLongProject/Assets/Scripts/Timeline/SetInvincible/SetInvincibleTrackMixerBehavior.cs
@@ -1,0 +1,15 @@
+using UnityEngine.Playables;
+
+namespace Timeline.SetInvincible
+{
+    /// <summary>
+    ///     Mixer behavior for hitbox track. Empty since hitboxes don't blend.
+    /// </summary>
+    public class SetInvincibleTrackMixerBehavior : PlayableBehaviour
+    {
+        public override void ProcessFrame(Playable playable, FrameData info, object playerData)
+        {
+            base.ProcessFrame(playable, info, playerData);
+        }
+    }
+}
